requirements.txt
httpx>=0.27,<1.0
tenacity>=8.2,<10.0
python-dotenv>=1.0,<2.0
pydantic>=2.5,<3.0
pandas>=2.1,<3.0
numpy>=1.26,<3.0
scipy>=1.11,<2.0
matplotlib>=3.8,<4.0
fredapi>=0.5,<1.0
rich>=13.0,<15.0
pytest>=8.0,<9.0
